refactor: class 폴더 분리
