+++ src/components/class/ClassDetail.jsx
