+++ src/components/classes/Class.jsx
@@ -16,8 +16,8 @@ const Class = ({
   const navigate = useNavigate();
 
   const clickHandler = () => {
+    dispatch(__getPrevChats({ class_id }));
     navigate(`/class/${redis_class_id}`);
-    dispatch(__getPrevChats());
   };
 
   return (
